1.20.x Port (#1132)
Port to 1.20.1 and Neoforged

Check TODOs:

Fonts are not working
Check block & entity gui's
Check pokemob model rendering
Repopulate creative tabs
Tabs only have one or two items to make sure it works properly
Check TODO comments

+++ src/main/java/pokecube/legends/worldgen/features/LargeUnrefinedAquamarine.java
@@ -258,7 +258,7 @@ public class LargeUnrefinedAquamarine extends Feature<LargeDripstoneConfiguratio
          {
             int i = this.originY - pos.getY();
             Vec3 vec3 = this.windSpeed.scale((double)i);
-            return pos.offset(vec3.x, 0.0D, vec3.z);
+            return pos.offset((int) vec3.x, 0, (int) vec3.z);
          }
       }
    }
